Added Profile Image Download function to the ScratchAPI
The GUI application now shows the profile image of the logged in user

Fixed some issues with the DownloadFollowers and DownloadFollowing options, where they would crash if certain files they are trying to write to were nonexistent

+++ ScratchDL/Data Types/LoginInfo.cs
@@ -8,7 +8,7 @@
     {
         public record class UserInfo
         (
-            int id,
+            long id,
             bool banned,
             string username,
             string token,
